1.3
Ok, so here's the changes adding in the two dimensional array for the
schedule. As you can see, it works like the other functions for
adding/removing classes. I've added drop downs for both choosing the
semester to schedule and remove a class. As of right now, this works.

The next step is adding in checks for both the semester a class is
available and it's prereqs. We'll have to do some string parsing for
these in the case of multiple semester availability and optional
prereqs.

We'll be focusing on checking the selected courses against the
"completed courses" array. obviously if it doesn't flag, we'll get an
alert. For removing courses from the schedule, we are also going to have
to check each individual course (that is left after initial course
removal) against the completed courses array. if we flag it, it gets
removed accordingly.

All of this will be done in the schedule class and remove scheduled
class functions.

Also, as a side note, and pretty important. the first position of each
array in the 2 dimensional scheduled courses array is simply the
semester name, so keep it in mind and make sure your code doesn't check
it by accident.